단어문제 문제 -> 유원우 디렉토리로 이동
